requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
python-dateutil==2.8.2
pytest==7.4.3
pytest-django==4.5.2